# docs/assets/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/hitl/queue": {
      "get": {
        "operationId": "HitlController_getQueue",
        "parameters": [
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filter by document status",
            "schema": {
              "default": "completed_ocr",
              "type": "string",
              "enum": [
                "completed_ocr",
                "all"
              ]
            }
          },
          {
            "name": "modelId",
            "required": false,
            "in": "query",
            "description": "Filter by model ID",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "maxConfidence",
            "required": false,
            "in": "query",
            "description": "Maximum confidence threshold (show fields below this)",
            "schema": {
              "default": 0.9,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Limit results",
            "schema": {
              "default": 50,
              "type": "number"
            }
          },
          {
            "name": "offset",
            "required": false,
            "in": "query",
            "description": "Offset for pagination",
            "schema": {
              "default": 0,
              "type": "number"
            }
          },
          {
            "name": "reviewStatus",
            "required": false,
            "in": "query",
            "description": "Filter by review status",
            "schema": {
              "$ref": "#/components/schemas/ReviewStatusFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of documents requiring human review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get review queue with filters",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/queue/stats": {
      "get": {
        "operationId": "HitlController_getQueueStats",
        "parameters": [
          {
            "name": "reviewStatus",
            "required": false,
            "in": "query",
            "description": "Filter by review status",
            "schema": {
              "$ref": "#/components/schemas/ReviewStatusFilter"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Queue statistics including total counts and average confidence",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueStatsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get queue statistics",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/sessions": {
      "post": {
        "operationId": "HitlController_startSession",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewSessionDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review session created with document and OCR data",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSessionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Start a review session",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/sessions/{id}": {
      "get": {
        "operationId": "HitlController_getSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review session with document, OCR data, and corrections",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewSessionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get review session details",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/sessions/{id}/corrections": {
      "post": {
        "operationId": "HitlController_submitCorrections",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitCorrectionsDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Corrections saved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitCorrectionsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Submit corrections for a session",
        "tags": [
          "hitl"
        ]
      },
      "get": {
        "operationId": "HitlController_getCorrections",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all corrections submitted for the session",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CorrectionsListResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get correction history for a session",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/sessions/{id}/submit": {
      "post": {
        "operationId": "HitlController_approveSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session approved and marked complete",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionActionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Approve and complete a review session",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/sessions/{id}/escalate": {
      "post": {
        "operationId": "HitlController_escalateSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EscalateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Session escalated for expert review",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionActionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Escalate a document for expert review",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/sessions/{id}/skip": {
      "post": {
        "operationId": "HitlController_skipSession",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "Session ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Session skipped",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionActionResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Skip a review session",
        "tags": [
          "hitl"
        ]
      }
    },
    "/api/hitl/analytics": {
      "get": {
        "operationId": "HitlController_getAnalytics",
        "parameters": [
          {
            "name": "startDate",
            "required": false,
            "in": "query",
            "description": "Start date for analytics period",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "required": false,
            "in": "query",
            "description": "End date for analytics period",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "reviewerId",
            "required": false,
            "in": "query",
            "description": "Filter by reviewer ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Review analytics including correction rates and session summaries",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalyticsResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Invalid API key"
          }
        },
        "security": [
          {
            "keycloak-sso": []
          },
          {
            "api-key": []
          }
        ],
        "summary": "Get HITL analytics",
        "tags": [
          "hitl"
        ]
      }
    },
...
  },
  "info": {
    "title": "Document Intelligence Platform API",
    "description": "Comprehensive REST API for document processing, OCR, workflow orchestration, model training, and human-in-the-loop review",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "keycloak-sso": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "api-key": {
        "type": "apiKey",
        "in": "header",
        "name": "x-api-key"
      }
    },
    "schemas": {
...
      "ReviewStatusFilter": {
        "type": "string",
        "enum": [
          "pending",
          "reviewed",
          "all"
        ]
      },
      "QueueDocumentOcrDto": {
        "type": "object",
        "properties": {
          "fields": {
            "type": "object",
            "description": "Extracted key-value pairs with confidence scores",
            "additionalProperties": true
          }
        },
        "required": [
          "fields"
        ]
      },
      "QueueSessionSummaryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "reviewer_id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "completed_at": {
            "format": "date-time",
            "type": "string"
          },
          "corrections_count": {
            "type": "number"
          }
        },
        "required": [
          "id",
          "reviewer_id",
          "status",
          "corrections_count"
        ]
      },
      "QueueDocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "original_filename": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "model_id": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          },
          "ocr_result": {
            "$ref": "#/components/schemas/QueueDocumentOcrDto"
          },
          "lastSession": {
            "$ref": "#/components/schemas/QueueSessionSummaryDto"
          }
        },
        "required": [
          "id",
          "original_filename",
          "status",
          "created_at",
          "updated_at",
          "ocr_result"
        ]
      },
      "QueueResponseDto": {
        "type": "object",
        "properties": {
          "documents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/QueueDocumentDto"
            }
          },
          "total": {
            "type": "number",
            "description": "Total number of documents matching the filter"
          }
        },
        "required": [
          "documents",
          "total"
        ]
      },
      "QueueStatsResponseDto": {
        "type": "object",
        "properties": {
          "totalDocuments": {
            "type": "number",
            "description": "Total documents in the review queue"
          },
          "requiresReview": {
            "type": "number",
            "description": "Documents requiring human review"
          },
          "averageConfidence": {
            "type": "number",
            "description": "Average confidence score across all documents"
          },
          "reviewedToday": {
            "type": "number",
            "description": "Number of documents reviewed today"
          }
        },
        "required": [
          "totalDocuments",
          "requiresReview",
          "averageConfidence",
          "reviewedToday"
        ]
      },
      "ReviewSessionDto": {
        "type": "object",
        "properties": {
          "documentId": {
            "type": "string",
            "description": "Document ID to review"
          },
          "minConfidence": {
            "type": "number",
            "description": "Minimum confidence threshold"
          }
        },
        "required": [
          "documentId"
        ]
      },
      "SessionDocumentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "original_filename": {
            "type": "string"
          },
          "storage_path": {
            "type": "string"
          },
          "ocr_result": {
            "$ref": "#/components/schemas/QueueDocumentOcrDto"
          }
        },
        "required": [
          "id",
          "original_filename",
          "storage_path",
          "ocr_result"
        ]
      },
      "ReviewSessionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "documentId": {
            "type": "string"
          },
          "reviewerId": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "startedAt": {
            "format": "date-time",
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "document": {
            "$ref": "#/components/schemas/SessionDocumentDto"
          },
          "corrections": {
            "type": "array",
            "description": "Corrections submitted during this session"
          }
        },
        "required": [
          "id",
          "documentId",
          "reviewerId",
          "status",
          "startedAt",
          "document"
        ]
      },
      "CorrectionDto": {
        "type": "object",
        "properties": {
          "field_key": {
            "type": "string",
            "description": "Field key being corrected"
          },
          "original_value": {
            "type": "string",
            "description": "Original extracted value"
          },
          "corrected_value": {
            "type": "string",
            "description": "Corrected value"
          },
          "original_conf": {
            "type": "number",
            "description": "Original confidence score"
          },
          "action": {
            "type": "string",
            "description": "Action taken on this field",
            "enum": [
              "confirmed",
              "corrected",
              "flagged",
              "deleted"
            ],
            "default": "confirmed"
          }
        },
        "required": [
          "field_key",
          "action"
        ]
      },
      "SubmitCorrectionsDto": {
        "type": "object",
        "properties": {
          "corrections": {
            "description": "Array of field corrections",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CorrectionDto"
            }
          }
        },
        "required": [
          "corrections"
        ]
      },
      "SubmitCorrectionsResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "corrections": {
            "type": "array",
            "description": "Saved correction records"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "sessionId",
          "corrections",
          "message"
        ]
      },
      "CorrectionRecordDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "fieldKey": {
            "type": "string"
          },
          "originalValue": {
            "type": "string"
          },
          "correctedValue": {
            "type": "string"
          },
          "originalConfidence": {
            "type": "number"
          },
          "action": {
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "fieldKey",
          "action",
          "createdAt"
        ]
      },
      "CorrectionsListResponseDto": {
        "type": "object",
        "properties": {
          "sessionId": {
            "type": "string"
          },
          "corrections": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CorrectionRecordDto"
            }
          }
        },
        "required": [
          "sessionId",
          "corrections"
        ]
      },
      "SessionActionResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "completedAt": {
            "format": "date-time",
            "type": "string"
          },
          "reason": {
            "type": "string",
            "description": "Escalation reason (escalate endpoint only)"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "status",
          "message"
        ]
      },
      "EscalateDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "description": "Reason for escalation"
          }
        },
        "required": [
          "reason"
        ]
      },
      "AnalyticsCorrectionsByActionDto": {
        "type": "object",
        "properties": {
          "confirmed": {
            "type": "number"
          },
          "corrected": {
            "type": "number"
          },
          "flagged": {
            "type": "number"
          }
        }
      },
      "AnalyticsSummaryDto": {
        "type": "object",
        "properties": {
          "totalSessions": {
            "type": "number"
          },
          "completedSessions": {
            "type": "number"
          },
          "totalCorrections": {
            "type": "number"
          },
          "confirmedFields": {
            "type": "number"
          },
          "correctedFields": {
            "type": "number"
          },
          "flaggedFields": {
            "type": "number"
          }
        },
        "required": [
          "totalSessions",
          "completedSessions",
          "totalCorrections",
          "confirmedFields",
          "correctedFields",
          "flaggedFields"
        ]
      },
      "AnalyticsResponseDto": {
        "type": "object",
        "properties": {
          "totalDocuments": {
            "type": "number"
          },
          "reviewedDocuments": {
            "type": "number"
          },
          "averageConfidence": {
            "type": "number"
          },
          "correctionRate": {
            "type": "number"
          },
          "correctionsByAction": {
            "$ref": "#/components/schemas/AnalyticsCorrectionsByActionDto"
          },
          "summary": {
            "$ref": "#/components/schemas/AnalyticsSummaryDto"
          }
        },
        "required": [
          "totalDocuments",
          "reviewedDocuments",
          "averageConfidence",
          "correctionRate",
          "correctionsByAction",
          "summary"
        ]
      },
...
    }
  }
}